Rename tests to match their implementation classes
